Cast predict scores to float before converting to numpy
Torch might throw a type error (see https://github.com/pytorch/pytorch/issues/109873) when a model uses a different precision (e.g., bf16). Casting the scores to float fixes this.

+++ sentence_transformers/cross_encoder/CrossEncoder.py
@@ -441,7 +441,7 @@ class CrossEncoder(PushToHubMixin):
         if convert_to_tensor:
             pred_scores = torch.stack(pred_scores)
         elif convert_to_numpy:
-            pred_scores = np.asarray([score.cpu().detach().numpy() for score in pred_scores])
+            pred_scores = np.asarray([score.cpu().detach().float().numpy() for score in pred_scores])
 
         if input_was_string:
             pred_scores = pred_scores[0]
